refactor: ♻️ consolidate member invite systems
